+++ packages/cli/src/controller/network/constants.ts
@@ -2,13 +2,13 @@
 // SPDX-License-Identifier: GPL-3.0
 
 import {JsonRpcProvider, Provider} from '@ethersproject/providers';
+import {formatUnits} from '@ethersproject/units';
 import {Wallet} from '@ethersproject/wallet';
 import {ProjectType, ContractSDK, networks} from '@subql/contract-sdk';
 import {GraphqlQueryClient} from '@subql/network-clients/dist/clients/queryClient';
 import {NETWORK_CONFIGS, SQNetworks, SQT_DECIMAL} from '@subql/network-config';
 import {base58Decode} from '@subql/utils';
 import {BigNumberish, ContractReceipt, ContractTransaction, Signer} from 'ethers';
-import {formatUnits} from 'ethers/lib/utils';
 import {z} from 'zod';
 import {Logger} from '../../adapters/utils';
 import {ProjectType as ProjectTypeGql} from './__graphql__/base-types';