+++ packages/cli/CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Ethers import error (#2984)
+
+### Changed
+- Update MCP SDK and lock down version (#2984)
 
 ## [6.6.1] - 2025-11-20
 ### Changed